chore: add tests to processRundown()
